+++ default.py
@@ -7,15 +7,13 @@ from __future__ import unicode_literals
 from __future__ import division
 
 import sys
-import argparse
 import logging
-import json
     
 # --- Kodi stuff ---
 import xbmcaddon
 
 # AKL main imports
-from akl import constants, settings
+from akl import constants, settings, addons
 from akl.utils import kodilogging, io, kodi
 from akl.scrapers import ScraperSettings, ScrapeStrategy
 
@@ -47,32 +45,20 @@ def run_plugin():
     for i in range(len(sys.argv)):
         logger.info('sys.argv[{}] "{}"'.format(i, sys.argv[i]))
     
-    parser = argparse.ArgumentParser(prog='script.akl.tgdbscraper')
-    parser.add_argument('--cmd', help="Command to execute", choices=['launch', 'scan', 'scrape', 'configure', 'update-settings'])
-    parser.add_argument('--type', help="Plugin type", choices=['LAUNCHER', 'SCANNER', 'SCRAPER'], default=constants.AddonType.LAUNCHER.name)
-    parser.add_argument('--server_host', type=str, help="Host")
-    parser.add_argument('--server_port', type=int, help="Port")
-    parser.add_argument('--rom_id', type=str, help="ROM ID")
-    parser.add_argument('--romcollection_id', type=str, help="ROM Collection ID")
-    parser.add_argument('--source_id', type=str, help="Source ID")
-    parser.add_argument('--entity_id', type=str, help="Entity ID")
-    parser.add_argument('--entity_type', type=int, help="Entity Type (ROM|ROMCOLLECTION|SOURCE)")
-    parser.add_argument('--akl_addon_id', type=str, help="Addon configuration ID")
-    parser.add_argument('--settings', type=json.loads, help="Specific run setting")
-    
+    parser = addons.AklAddonArguments('script.akl.tgdbscraper')
     try:
-        args = parser.parse_args()
+        parser.parse()
     except Exception as ex:
         logger.error('Exception in plugin', exc_info=ex)
-        kodi.dialog_OK(text=parser.usage)
+        kodi.dialog_OK(text=parser.get_usage())
         return
         
-    if args.type == constants.AddonType.SCRAPER.name and args.cmd == 'scrape':
-        run_scraper(args)
-    elif args.cmd == "update-settings":
+    if parser.get_command() == addons.AklAddonArguments.SCRAPE:
+        run_scraper(parser)
+    elif parser.parser.cmd == "update-settings":
         update_plugin_settings()
     else:
-        kodi.dialog_OK(text=parser.format_help())
+        kodi.dialog_OK(text=parser.get_help())
         
     logger.debug('Advanced Kodi Launcher Plugin: TGDB Scraper -> exit')
 
@@ -80,31 +66,34 @@ def run_plugin():
 # ---------------------------------------------------------------------------------------------
 # Scraper methods.
 # ---------------------------------------------------------------------------------------------
-def run_scraper(args):
+def run_scraper(args: addons.AklAddonArguments):
     logger.debug('========== run_scraper() BEGIN ==================================================')
     pdialog = kodi.ProgressDialog()
     
-    settings = ScraperSettings.from_settings_dict(args.settings)
+    settings = ScraperSettings.from_settings_dict(args.get_settings())
     scraper_strategy = ScrapeStrategy(
-        args.server_host,
-        args.server_port,
+        args.get_webserver_host(),
+        args.get_webserver_port(),
         settings,
         TheGamesDB(),
         pdialog)
     
-    if args.entity_type == constants.OBJ_ROM:
+    if args.get_entity_type() == constants.OBJ_ROM:
         logger.debug("Single ROM processing")
-        scraped_rom = scraper_strategy.process_single_rom(args.entity_id)
+        scraped_rom = scraper_strategy.process_single_rom(args.get_entity_id())
         pdialog.endProgress()
         pdialog.startProgress('Saving ROM in database ...')
-        scraper_strategy.store_scraped_rom(args.akl_addon_id, args.entity_id, scraped_rom)
+        scraper_strategy.store_scraped_rom(args.get_akl_addon_id(), args.get_entity_id(), scraped_rom)
         pdialog.endProgress()
     else:
         logger.debug("Multiple ROM processing")
-        scraped_roms = scraper_strategy.process_roms(args.entity_type, args.entity_id)
+        scraped_roms = scraper_strategy.process_roms(args.get_entity_type(), args.get_entity_id())
         pdialog.endProgress()
         pdialog.startProgress('Saving ROMs in database ...')
-        scraper_strategy.store_scraped_roms(args.akl_addon_id, args.entity_type, args.entity_id, scraped_roms)
+        scraper_strategy.store_scraped_roms(args.get_akl_addon_id(),
+                                            args.get_entity_type(),
+                                            args.get_entity_id(),
+                                            scraped_roms)
         pdialog.endProgress()